Rename .hpp to .h
